user_service/requirements.txt
asgiref==3.8.1
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
sqlparse==0.5.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
drf-yasg